Include all years data for each tile
Because tile attributes may be used for charting, it seems like we would
want to make them have access to all data years available for a given
geography. An example would be if someone clicks on a county in 2010, we
would want to create a line chart for that county's entire history,
rather than just 2010-present or pull it from a separate source.